Fix scoping issues with new namespace
Scoping was achieved before with a nesting of namespaces.
This was changed in the last commit so that I can keep the pc in the
namespace and mutate it when needed, it was too hard with the nested
namespaces. The scoping issue was fixed by putting a 'scope-separator'
in the namespace when a new scope is introduced. Variables can be
shadowed when 'in' a new scope but not when in the same scope.

Procedures now introduce a new scope