instructions are now created with macros!!
Created a macro to make instructions. This allows for me to generate a
function for each instruction. This means instructions are not special
anymore, and can be 'applied' like macros or other builtin functions.

The syntax for the macro is an association list with the key being a
list of types that the instruction requires and the value follows the
syntax found here:
http://devernay.free.fr/hacks/chip8/C8TECH10.HTM#3.0, but with spaces
between each part.