Changed handling of v registers. I think this is a good change
V registers are now a part of the namespace and evaluate to a list
containing 'v' and a number. By default this list evaluates to itself
but when is later transformed into just the number when needed.

I also rewrote the chip8-eval function. I think it looks nicer now.