Implemented dereference operator, among other things
Dereference operator implemented. This operator assumes the number
given to it is a valid location in memory and finds the byte at the
location. Currently, it works by looping over what should be a fully
compiled program except for '@' operators and resolving those. I can't
do it during the normal evaluation because I only save the state of
the binary at the very end.

I changed the implementation of 'include'. It now does not pad the
numbers with a zero if there are an odd number of arguments.

I don't give an error on undefined functions. This may change.

Changed the 'lab' keyword to ':', which is what I originally wanted,
but didn't know how to use special characters at the time.

I also fixed the arithmetical functions. I need to ignore the env
parameter that is given to all functions/macros so I made a little
macro to make things look nicer.