I decided to move most of the builtin stuff to chip8-eval
I thought it was nicer, maybe.
I also restructured the whole program.

Also, the program counter will remain as a variable for now, because
when it is in the namespace only one instance of pc is ever created,
which is good. This does not happen if I put it in the struct.